pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "autovram"
description = "GPU Auto-Optimizer: automatically finds the fastest stable batch/precision configuration for ML scripts and vLLM."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "autovram contributors" }]
keywords = ["gpu", "cuda", "vram", "autotune", "pytorch", "vllm"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
]

dynamic = ["version"]

[project.optional-dependencies]
yaml = ["PyYAML>=6.0.1"]
vllm = ["vllm>=0.6.0"]
llm-autobatch = ["llm-autobatch"]

[project.scripts]
autovram = "autovram.cli.app:app"

[project.urls]
Homepage = "https://github.com/your-org/autovram"
Repository = "https://github.com/your-org/autovram"
Issues = "https://github.com/your-org/autovram/issues"

[tool.hatch.version]
path = "src/autovram/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/autovram"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pre-commit>=3.6.0",
  "ruff>=0.6.0",
  "mypy>=1.8.0",
  "types-PyYAML; extra == 'yaml'",
]
